move dirSize -> getDirSize
