+++ src/main/java/org/broadinstitute/hellbender/engine/AbstractConcordanceWalker.java
@@ -13,9 +13,10 @@ import org.broadinstitute.hellbender.exceptions.GATKException;
 import org.broadinstitute.hellbender.tools.walkers.validation.ConcordanceState;
 import org.broadinstitute.hellbender.utils.SimpleInterval;
 import org.broadinstitute.hellbender.utils.Utils;
-import org.broadinstitute.hellbender.utils.gcs.BucketUtils;
+import org.broadinstitute.hellbender.utils.io.IOUtils;
 
 import java.io.File;
+import java.nio.file.Files;
 import java.util.Iterator;
 import java.util.Optional;
 import java.util.Spliterator;
@@ -72,7 +73,7 @@ public abstract class AbstractConcordanceWalker extends GATKTool {
     }
 
     private void initializeTruthVariantsIfNecessary() {
-        if (! BucketUtils.fileExists(truthVariantsFile)) {
+        if (!Files.exists(IOUtils.getPath(truthVariantsFile))) {
             throw new IllegalArgumentException("Truth variants file " + truthVariantsFile + " does not exist or is not readable.");
         }
