+++ src/main/java/org/broadinstitute/hellbender/tools/spark/pathseq/PSUtils.java
@@ -5,7 +5,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.spark.api.java.JavaRDD;
 import org.broadinstitute.hellbender.utils.Utils;
-import org.broadinstitute.hellbender.utils.gcs.BucketUtils;
+import org.broadinstitute.hellbender.utils.io.IOUtils;
 import org.broadinstitute.hellbender.utils.read.GATKRead;
 
 import java.util.Collection;
@@ -62,7 +62,7 @@ public final class PSUtils {
         if (numReducers != 0) {
             return numReducers;
         }
-        return 1 + (int) (BucketUtils.dirSize(inputPath) / targetPartitionSize);
+        return 1 + (int) (IOUtils.getDirSize(inputPath) / targetPartitionSize);
     }
 
     /**
