+++ src/main/java/org/broadinstitute/hellbender/tools/spark/sv/evidence/FindBadGenomicKmersSpark.java
@@ -20,12 +20,12 @@ import org.broadinstitute.hellbender.tools.spark.sv.StructuralVariationDiscovery
 import org.broadinstitute.hellbender.tools.spark.sv.utils.*;
 import org.broadinstitute.hellbender.tools.spark.utils.HopscotchMap;
 import org.broadinstitute.hellbender.utils.Utils;
-import org.broadinstitute.hellbender.utils.gcs.BucketUtils;
 import scala.Tuple2;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -184,7 +184,7 @@ public final class FindBadGenomicKmersSpark extends GATKSparkTool {
                                                          final String fastaFilename) {
         try ( BufferedReader rdr = new BufferedReader(new InputStreamReader(IOUtils.openInputStream(
                 IOUtils.getPath(fastaFilename)))) ) {
-            final List<SVKmer> kmers = new ArrayList<>((int) BucketUtils.fileSize(fastaFilename));
+            final List<SVKmer> kmers = new ArrayList<>((int) Files.size(IOUtils.getPath(fastaFilename)));
             String line;
             final StringBuilder sb = new StringBuilder();
             final SVKmer kmerSeed = new SVKmerLong();
